Fix hitung dan sorting kontrabon

+++ internal/domain/kontrabon/kontrabon.go
@@ -4,12 +4,12 @@ const STATUS_CREATED = "CREATED"
 const STATUS_LUNAS = "LUNAS"
 
 type Kontrabon struct {
-	ID          string `json:"id"`
-	Code        string `json:"code"`
-	CreatedTime string `json:"createdTime"`
-	Status      string `json:"status"`
-	Total       int64  `json:"total"`
-	CustomerID  string `json:"customerId"`
+	ID          string  `json:"id"`
+	Code        string  `json:"code"`
+	CreatedTime string  `json:"createdTime"`
+	Status      string  `json:"status"`
+	Total       float64 `json:"total"`
+	CustomerID  string  `json:"customerId"`
 }
 
 type CreateRequest struct {
